+++ CHANGELOG.md
@@ -7,6 +7,27 @@ The format is based on [keep a changelog](http://keepachangelog.com) and this pr
 
 ## [Unreleased]
 
+## [1.2.0] - 2023-08-29
+### Added
+- Add server interface for Stats gameplay system.
+
+### Changed
+- Pin dependencies to compatible versions of Nakama common at v1.28.1.
+- Return all Reward Tiers when an Event Leaderboard is fetched for the current user.
+
+### Fixed
+- Fix weighted reward error when definition is empty (instead of nil).
+
+## [1.1.0] - 2023-08-23
+### Added
+- Add server interface for Event Leaderboards gameplay system.
+
+## [1.0.4] - 2023-08-22
+### Changed
+- Add ChannelMessageAck message to proto definition.
+
+### Fixed
+- Expose server functions for reward and roll in Hiro. 
 
 ## [1.0.3] - 2023-08-10
 ### Added